Simplify setup.py; remove unnecessary py_modules
The python-tblib is a Python package (multiple files) and not simply a
module (single file). This is correctly specified in setup.py by the
packages directive. There is no need to also list each individual file
as a Python module, the packages directive already handles that. For
additional information on these commands, see:

https://docs.python.org/3/distutils/setupscript.html#listing-individual-modules